+++ test/unit/fuzz-tools.test.js
@@ -307,6 +307,7 @@ import * as liRadsV327 from '../../lib/li-rads-v327.js';
 import * as montrealIbdV328 from '../../lib/montreal-ibd-v328.js';
 import * as parisClassificationV329 from '../../lib/paris-classification-v329.js';
 import * as nottinghamNpiV330 from '../../lib/nottingham-npi-v330.js';
+import * as fitzpatrickV331 from '../../lib/fitzpatrick-v331.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -559,6 +560,7 @@ const MODULES = {
   'montreal-ibd-v328.js': montrealIbdV328,
   'paris-classification-v329.js': parisClassificationV329,
   'nottingham-npi-v330.js': nottinghamNpiV330,
+  'fitzpatrick-v331.js': fitzpatrickV331,
   'tb-testing.js': tbTesting,
 };